Provide the parasite_python API whether Python is enabled or not.
We also blacklist Python support for Gimp, since Gimp is very broken
with regards to its GtkScaleButton widget. It actually registers a
duplicate "GtkScaleButton," which breaks horribly as soon as pygtk tries
to also register GtkScaleButton. This needs to be fixed in Gimp for
Python support to work at all.