Add a notice and link about GtkInspector.
Parasite hasn't been in development for years, but sometimes people find
it and want to use it. The modern-day equivalent is GtkInspector, which
was based on Parasite, but is actively maintained.

This change adds a notice to the top of the page, mentioning that the
project is dead, pointing to GtkInspector, and telling people what we're
up to now.